Add Choreo logo
